+++ CHANGELOG.md
@@ -44,6 +44,7 @@ Please make sure to add your changes to the appropriate categories:
 - Added `TimedEnvelope<T>` dt-aware follower
 - Added `AnalogBiquad<T>` s-domain section
 - Added `butterworth_lowpass` normalized prototype
+- Added `lp_to_lp`/`lp_to_hp` s-domain transforms
 
 ### Changed